Revised the definiton of seed size: (TO:0000391):  A seed anatomy and morphology trait (TO:0000184) which is associated with the size of a seed (PO:0009010).
Revised comment: Seed size is determined by the ratio of the seed length to
seed width and is a component of seed quality (TO:0000162). Seed size may
be a qualitative measure or may be organized into categories ranging from
small to medium to large. Refers to size (PATO:0000117): A morphology
quality inhering in a bearer by virtue of the bearer's physical magnitude.

added the relationship to seed quality as a cross product:

is_a: TO:0000184 ! seed anatomy and morphology trait
intersection_of: TO:0000184 ! seed anatomy and morphology trait
intersection_of: part_of TO:0000162 ! seed quality

Fixed the dbxref to SF: from OBO_SF_ID to  OBO_SF_TO.  Fixed dbxref to size PATO:0000117, instead of PATO: 0000162.