spikelet anatomy and morphology trait (TO:0000657): A reproductive shoot system anatomy and morphology trait (TO:0000786) which is associated with changes in the anatomy or structure of the spikelet (PO:0009051) and its parts.
Moved to be  is_a inflorescence anatomy and morphology trait (TO:0000373)- revised def'n: An inflorescence anatomy and morphology trait (TO:0000373) which is associated with a spikelet (PO:0009051).

Updated SF tracker #143 and added reference to it.

flower anatomy and morphology trait (TO:0000499): An inflorescence anatomy and morphology trait (TO:0000373) which is associated with a flower (PO:0009046).

Moved to be  is_a reproductive shoot system anatomy and morphology trait (TO:0000786) 
revised def'n:   A reproductive shoot system anatomy and morphology trait (TO:0000786) which is associated with a flower (PO:0009046).

Updated SF tracker #56 and added reference to it.


Added missing dbxref:  OBO_SF2_TO: 56 to shoot system anatomy and morphology trait (TO:0000077); reproductive shoot system anatomy and morphology trait (TO:0000786);   flower anatomy and morphology trait (TO:0000499)

Created new term: infructescence anatomy and morphology trait (TO:0000920): A reproductive shoot system anatomy and morphology trait (TO:0000786) which is associated with an infructescence (PO:0006342).

Currently: 
- ear diameter (TO:0000433): Diameter measurement of a corn ear. is_a inflorescence diameter (TO:0000679).
- ear circumference (TO:0000681):  undefined
- ear number (TO:0000443): Number of ears found in a plant or and average number present in population.
- ear length (TO:0000431): undefined, grayed out? 

Moved all to new term: 
- infructescence anatomy and morphology trait (TO:0000920): A reproductive shoot system anatomy and morphology trait (TO:0000786) which is associated with an infructescence (PO:0006342).

created new term: 
infructescence axis anatomy and morphology trait (new TO:0000921): A shoot axis anatomy and morphology trait (TO:0000739) which is associated with an infructescence axis (PO:0025242).

Moved existing term 'cob diameter' (TO:0000678) to be is_a infructescence axis anatomy and morphology trait (new TO:0000921).   Def'n: An infructescence axis anatomy and morphology trait (TO:0000921) associated with the diameter of a cob.   Was previously undefined.
Added ref to both for OBO_SF2_TO: #243