Revised def'ns: Referred to PATO terms and new ear infructescence (PO:0025597) PO term for maize ear. * ear length (TO:0000431): An infructescence anatomy and morphology trait (TO:0000920) that is associated with the length of an ear infructescence (PO:0025597) from base to tip. Comment: Refers to length (PATO:0000122): A 1-D extent quality which is equal to the distance between two points. In maize, this is measured in centimeters.
* ear diameter (TO:0000433):  An infructescence anatomy and morphology trait (TO:0000920) that is associated with the diameter of an ear infructescence (PO:0025597).
An infructescence anatomy and morphology trait (TO:0000920) that is associated with the diameter of an ear infructescence (PO:0025597). 
Comment: Refers to diameter (PATO:0001334): A length quality which is equal to the length of any straight line segment that passes through the center of a circle and whose endpoints are on the circular boundary. In maize, this is measured in centimeters.

* new term: ear weight (TO:0000923): An infructescence anatomy and morphology trait (TO:0000920) that is associated with the weight (mass) of an ear infructescence (PO:0025597). 
Comment: Refers to mass (PATO:0000125): A physical quality that inheres in a bearer by virtue of the proportion of the bearer's amount of matter.

Added all three to the maize/corn (Zea mays) subset, added dbxrefs to MaizeGDB #s ?? as  placeholders.

Moved ear number (TO:0000443) to be is_an 'above ground biomass yield trait' (TO:0000327).  Currently, we do not have any an infrutescence yield trait term, but this could be added as an is_a parent.  Need to work on definition.
Made average grain weight (TO:0000589) a fruit yield trait (TO:0000871).