- Merged the older, undocumented term: inflorescence diameter (TO:0000679) into inflorescence diameter (TO:0000802). removed the is_a relation to inflorescence size (TO:0006031). Added ref to OBO_SF2_PO:96
inflorescence weight (TO:0000583)- Average weight of the inflorescence carrying the mature fruits from the plants or tillers in a given study.  This is an infructescence, not and inflorescence. 

Was is_a yield trait (TO:0000371), moved to be is_a infructescence anatomy and morphology trait (TO:0000920).  Could also go in the yield category once we work on it some more.

Revised name and definition: 
infructescence weight (TO:0000583):  An infructescence anatomy and morphology trait (TO:0000920) that is associated with  the average weight of an infructescence bearing mature fruit (PO:0009001).

Made ear weight (TO:0000923): a child of the revised term: 
An infructescence weight (TO:0000583) trait that is associated with the weight (mass) of an ear infructescence (PO:0025597).

We have duplicates for the 'inflorescence width' term:

inflorescence width (TO:0000637): undefined, no dbxrefs; no child terms; is_a inflorescence size (TO:0006031), a few direct annotations.

inflorescence width (TO:0000804): An inflorescence anatomy and morphology trait (TO:0000373) which is associated with the width of an inflorescence (PO:0009049).

is_a inflorescence anatomy and morphology trait (TO:0000373).



Merged the older, undocumented term: inflorescence width (TO:0000637) into inflorescence width (TO:0000804). removed the is_a relation to inflorescence size (TO:0006031). 

Made inflorescence width (TO:0000804) part_of inflorescence size (TO:0006031).