Added def'n for existing term that had no definition:
ear circumference (TO:0000681): An infructescence anatomy and morphology trait (TO:0000920) that is associated with the circumference of an ear infructescence (PO:0025597).
comment: Refers to circumference (PATO:0001648): An 1-D extent quality which is equal to the length of the closed curve of a circle. Ear circumference (TO:0000681) is equal to ear diameter (TO:0000433) times 3.14.

Merged duplicated terms:  total water soluble content (TO:0002711) into total water soluble content (TO:0020088)

total water soluble content (TO:0002711) : "The concentration of water soluble components in the plant as determined by refractive index." [GR:pj]
 
total water soluble content (TO:0020088):  "A trait to measure the concentration of water soluble components in the plant as determined by refractive index." 
comment: This trait is usually highly correlated with total soluble sugar content (TO:0000340). Often determined in fresh tissue.

Revised def'n: total water soluble content (TO:0002711): A metabolite content related trait (TO:0000281) associated with the concentration of water soluble components in a plant, as determined by refractive index.
is_a: TO:0000281 ! metabolite content related trait

Removed synonym: "total dissolved solids content" EXACT as it was causing an error.