Revised definition: A growth and development trait (TO:0000357) that has as primary participant a plant anatomical entity (PO:0025131) or a constituent cellular component (GO:0005575) contained therein. OBO_SF2_TO: 248
Moved existing def'n to comment: Growth and development traits are associated with the development of the whole plant or plant parts.

Added new terms: 

plant structure growth and development trait (TO:0000928):  A growth and development trait (TO:0000357) that has as primary participant a plant structure (PO:0009011).
OBO_SF2_TO: 252

multi-tissue plant structure growth and development trait (TO:0000926): A plant structure growth and development trait (TO:0000928) that has as primary participant a multi-tissue plant structure (PO:0025496).
OBO_SF2_TO: 250

plant organ growth and development trait (TO:0000927):
A multi-tissue plant structure development trait (TO:0000927) that has as primary participant a plant organ (PO:0009008).
OBO_SF2_TO:251

leaf growth and development trait (TO:0000655): 
Current name and Def'n: leaf development trait  (TO:0000655):
Traits associated with observing effects on the development of the leaf and its parts.

 Currently: leaf development trait (TO:0000655) is a child of shoot development trait (TO:0000654).  
Moved to be is_a plant organ development trait (TO:0000927).

Revised definition: A plant organ development trait (TO:0000927) that has as primary participant a leaf (PO:0025034).
OBO_SF2_TO:253


leaf shattering (TO:0000108):  Currently a child of leaf anatomy and morphology trait (TO:0000748). Moved to be is_a leaf growth and development trait (TO:0000655).

revised definition:  A leaf growth and development trait (TO:0000655) which is associated with the extent of leaf shattering based on the development of a leaf abscission zone (PO:0006501).

Removed reference to PATO:0000074, as it has been obsoleted.  
Added dbxref to 'leaf abscission' GO:0060866  def'n: The controlled shedding of a leaf.


Currently, seed development trait (TO:0000653) is a direct child of growth and development trait (TO:0000357).
Deff'n: Traits associated with observing effects on the development of the seed and its parts.

Revised: seed growth and development trait (TO:0000653): A multi-tissue plant structure development trait (TO:0000927) that has as primary participant a seed (PO:0009010).
OBO_SF2_TO:254

New term: fruit growth and development trait (TO:0000929): A multi-tissue plant structure development trait (TO:0000927) that has as primary participant a fruit (PO:0009001).
OBO_SF2_TO:255


New term: fruit growth and development trait (TO:0000929): A multi-tissue plant structure development trait (TO:0000927) that has as primary participant a fruit (PO:0009001).

Added new mid-level term to better reflect the hierarchy of the anatomy and morphology trait (TO:0000017) branch and the the Plant Ontology. This will serve as the parent terms for the fruit filling and ripening traits.


fruit ripening trait (TO:0002633)- made it a child of the fruit growth and development trait (new TO:0000929)

Revised def'n:  A fruit growth and development trait (TO:0000929) that is associated with the variation in the extent of fruit ripening (GO:0009835).

Comment:  Fruit ripening (GO:0009835) is an aging process that has as participant a fruit. Ripening causes changes in one or more characteristics of a fruit (color, aroma, flavor, texture, hardness, cell wall structure) and may make it more attractive to animals and aid in seed dispersal.
OBO_SF2_TO:256