Revised definitions: quality trait (TO:0000597): A plant trait (TO:0000387) that is often associated with the commercial and /or economical value of the plant product, or its overall improvement.
Added new Comment: Quality traits are often also classified as belonging to another upper level class such as: anatomy and morphology trait (TO:0000017), biochemical trait (TO:0000277), growth and development trait (TO:0000357), stature or vigor trait (TO:0000133) and so on.  Created new OBO_SF_TO: 260

Renamed seed quality (TO:0000162):  grain quality trait (TO:0000162), moved it to be a child of fruit quality trait (TO:0002728)
Revised def'n:  A fruit quality trait (TO:0002728) that is associated with the commercial and /or economical value of the grain (i.e. fruit (PO:0009001)), or its overall improvement.   Added missing dbxref:  OBO_SF_TO: 208

Revised definition:  fruit quality trait (TO:0002728): A quality trait (TO:0000597) associated with the commercial and /or economical value of a fruit (PO:0009001) , or its overall improvement. OBO_SF_TO: 263

seed density TO:0000612- renamed grain density trait
Revised def'n:A grain quality trait (TO:0000162) that is a measure of the density or weight per unit of volume of a grain at a standardized moisture level.
OBO_SF_TO: 262

Created new term: seed quality trait (TO:0000931): A quality trait (TO:0000597) associated with the commercial and /or economical value of a seed (PO:0009010) , or its overall improvement.   OBO_SF_TO: 209

Moved the following terms to be is_a seed quality trait (TO:0000931), Most of these names and all these definitions need to be revised:
- endosperm quality (TO:0000587)
- cooking or brewing quality (TO:0000667)
- DSDS50 (TO:0006069)
- germinability at low temperature (TO:0000483)
- germinability in dark (TO:0006068)
- seed coat color TO:0000190
- seed composition based quality trait TO:0000488- no def'n
- seed fertility TO:0000639- no def'n
- seed size TO:0000391
- seed viability  TO:0000345 - no def'n
- water absorption by seed TO:0002648
- wrinkled seed TO:0000718 - no def'n

Revised def'n: 
grain thickness (TO:0000399):  A grain quality trait (TO:0000399) which is associated with the thickness of a dehulled grain with a pericarp (PO:0009084).
Moved to be is_a grain quality trait (TO:0000399)
part of grain size: TO:0000397
OBO_SF_TO: 264