Rmoved is_a relationship between the following terms and inflorescence development trait (TO:0000621) 	- spikelet number (TO:0000456): was both is_a spikelet anatomy and morphology trait (TO:0000657), and also an inflorescence development trait (TO:0000621)
- panicle type (TO:0000089): was both  is_a panicle anatomy and morphology trait (TO:0000847) , and also an inflorescence development trait (TO:0000621)

- created new term: 
plant phenological trait (new; TO:0000933):
A growth and development trait (TO:0000357) that is concerned with the timing of the occurrence of plant structure development stages (PO:0009012) and the relevant biological processes (GO:0008150).

revised definition and parentage: 
flowering time (TO:0002616):flowering time (TO:0002616): A plant phenological trait (TO:0000933) associated with the timing of anthesis (flowering stage; PO:0007616).

related synonyms: anthesis timing; time of flowering

Revised name and def'n:
Was: shoot development trait (TO:0000654): Traits associated with observing effects on the development of the shoot and its parts.

Revised
shoot system growth and development trait (TO:0000654): A plant growth and development trait (TO:0000357) that is associated with the growth of the shoot system (PO:0009006), and shoot system development stages (PO:0025527).

Other changes - all these need trackers and need to have the definitions revised and updated: 
Direct children of plant growth and development trait (TO:0000357) were moved to the following classes:
- organ identity (TO:0006018), homeotic development trait (TO:0002611) and root development trait (PO:0000656) > plant organ growth and development trait (TO:0000927)
- gynoecium development trait (TO:0020110)  and embryo development trait (TO:0000620) > plant structure growth and development trait (TO:0000928)

reproductive growth time (TO:0000366), vegetative growth time (TO:0000369), days to inflorescence exsertion (TO:0001021), life cycle habit (TO:0002725), DSDS50 (TO:0006069) and days to germinate (TO:0006061) >  plant phenological trait (new; TO:0000933)

sporocyte development trait (TO:0000722)  and cell organelle development trait (TO:0002755) > cell growth and development trait (TO:0002686) > plant structure growth and development trait (TO:0000928)

vascular bundle development trait (TO:0020109) >  tissue development trait (TO:0006015) > plant structure growth and development trait (TO:0000928)

germinability in dark (TO:0006068) >   seed growth and development trait (TO:0000653)