-Moved average grain weight (TO:0000589) to be is a grain yield trait (TO:0000396). revised def'n: A grain yield trait (TO:0000396) associated with the average weight of a grain, including the caryopsis hull (PO:0006000). Moved to comment: The average grain weight (TO:0000589) can be used to calculate the overall yield or produce from a field.
- Moved grain yield per panicle (TO:0000450) to is a grain yield trait (TO:0000396)

- Moved grain yield per plant (TO:0000449) to is a grain yield trait (TO:0000396)

Created new term: 
infructescence yield trait (TO:0000936):
An above ground biomass yield trait (TO:0000327) associated with the characteristics of an infructescence (PO:0006342), or the number born on a plant.
*Needs a SF tracker

Moved existing terms to be is a children of infructescence yield trait (TO:0000936):
- ear number (TO:0000443):  
Revised def'n: An infructescence yield trait (TO:0000936) that is associated with the number of ears (ear infructescence; PO:0025597) on a plant or the average number present in population.
*OBO_SF_TO:  244

- kernel row length (TO:0000688): , kernel row number (TO:0000687): 

Created new term: 
kernel number per ear (TO:0000937): An infructescence yield trait (TO:0000936)) that is associated with the number of kernels (fruit; PO:0009001) per ear infructescence (PO:0025597).
*OBO_SF_TO: 284
MaizeGDB_QTL: 99383
PMID: 22027895

Created new term: 
kernel number per row
An infructescence yield trait (TO:0000936) that is associated with the number of kernels (fruit; PO:0009001) per row on an ear infructescence (PO:0025597).
cd OBo	: 78114
*OBO_SF_TO: 285