above ground biomass yield trait (TO:0000327): A yield trait (TO: 0000371) that is associated with the amount of a plant anatomical entity (PO:0025131) which is produced above ground.
revised name and def'n: shoot system yield trait (TO:0000327): A yield trait (TO: 0000371)  that is associated with the amount of a plant structure (PO:0009011) or portion of plant substance (PO:0025161) produced by the shoot system (PO:0009006) which is harvested.
Can add intermediate levels terms as needed.
Added missing Xref to SF: OBO_SF2_TO: 290 

Child terms need further revisions. Revised infructescence yield trait (TO:0000936), fruit yield trait (TO:0000871), seed yield trait (TO:0000905), and leaf yield trait (TO:0000870) to reflect new parent name. 

- seed yield trait (TO:0000905)- created has part relations to: seed weight (TO:0000181); seed size (TO:0000391); and seed number (TO:0000445).

- Moved 1000-seed weight (TO:0000382) and 100 seed weight (TO:0000269) to be children of seed weight (TO:0000181), rather than fruit weight (TO:0002746). Need further revisions.

- Moved culm number (TO:0000027) to be is_a shoot system anatomy and morphology trait (TO:0000077) rather than a growth and development trait. Needs further revisions.

seed number (TO:0000445): moved it to be a fruit anatomy and morphology trait (TO:0002629).  Revised definition: A fruit anatomy and morphology trait (TO:0002629) which is associated with the number of seeds (PO:0009010) in a fruit (PO:0009001).

fruit number trait (TO:0000894): moved to be an an infructescence anatomy and morphology trait (TO:0000920); revised def'n to reflect.

Added has_part relations from fruit yield trait (TO:0000871) to: fruit number trait (TO:0000894), fruit size (TO:0002625) and fruit weight (TO:0002746).

-Removed has–part relation between grain quality trait (TO:0000162) and seed weight (TO:0000181), seed (TO:0000486) and seed shape (TO:0000484).


Removed comment from seed number (TO:0000445):
It is directly related to the filled grain number (TO:0000447) and grain number (TO:0002759).