Created new terms: Need to add xrefs to the GitHub trackers once  they are created.
aleurone anatomy and morphology trait (TO:0000942): A portion of plant tissue anatomy and morphology trait (TO:0000843) which is associated with the qualities of the aleurone layer (PO:0005360).
is_a  portion of plant tissue anatomy and morphology trait (TO:0000843)
part_of endosperm anatomy and morphology trait (TO:0000575)

aleurone layer appearance (TO:0000944): An aleurone layer anatomy and morphology trait (TO:0000942) that is associated with the appearance of the aleurone layer (PO:0005360).

aleurone layer color (TO:0000943): An aleurone layer appearance (TO:0000944) that is associated with the color of the aleurone layer (PO:0005360).