Created new term: pericarp anatomy and morphology trait (TO:0000945): A portion of plant tissue anatomy and morphology trait (TO:0000843) which is associated with the qualities of the pericarp (PO:0009084).
Moved pericarp color (TO:0000707) to be a child of pericarp anatomy and morphology trait (TO:0000945.
 Revised def'n: was: Describes the color of fruit wall (pericarp).
New def'n: A pericarp anatomy and morphology trait (TO:0000945) that is associated with the color of the pericarp (PO:0009084).
Added Xref to color PATO:0000014
Added part_of relation to fruit color (TO:0002617) and secondary is_a relation to fruit quality trait (TO:0002728).
Need to come back and look at the synonyms.

Existing fruit quality trait (TO:0002728): pericarp texture (TO:0002623)- no def'n etc
Added new is_a parent:   pericarp anatomy and morphology trait (TO:0000945).
pericarp texture (TO:0002623): A pericarp anatomy and morphology trait (TO:0000945) that is associated with the texture of the pericarp (PO:0009084).
Added Xref to texture PATO:0000150