portion of plant tissue anatomy and morph traits: carp edition!
added term: endocarp anatomy and morphology trait (TO:0000946): A portion of plant tissue anatomy and morphology trait (TO:0000843) which is associated with the qualities of the endocarp (PO:0009086).
added relationship: endocarp color (TO:0002618) is_a endocarp anatomy and morphology trait (TO:0000946)
removed relationship:  endocarp color (TO:0002618) is_a pericarp color(TO:0000707)

added term: epicarp anatomy and morphology trait (TO:0000947): A portion of plant tissue anatomy and morphology trait (TO:0000843) which is associated with the qualities of the epicarp (PO:xxx).
added relationship: epicarp color (TO:0002620) is_a epicarp anatomy and morphology trait (TO:0000947)
removed relationship: epicarp color (TO:0002620) is_a pericarp color(TO:0000707)

added term: mesocarp anatomy and morphology trait (TO:0000948): A portion of plant tissue anatomy and morphology trait (TO:0000843) which is associated with the qualities of the mesocarp (PO:0009087).
added relationship: mesocarp color (TO:0002619) is_a mesocarp anatomy and morphology trait (TO:0000948)
removed relationship: mesocarp color (TO:0002619) is_a pericarp color(TO:0000707)

added relationships:
endocarp color (TO:0002618) has_part fruit color(TO:0002617)
epicarp color (TO:0002620) has_part fruit color(TO:0002617)
mesocarp color (TO:0002619) has_part fruit color(TO:0002617)

epicarp color (TO:0002620)
mesocarp color (TO:0002619)

portion of plant tissue anatomy and morphology trait (TO:0000843)

previously without def'n.  Added definitions to:
endocarp color (TO:0002618): An endocarp anatomy and morphology trait (TO:0000946) that is associated with the color of the endocarp (PO:0009084).
epicarp color (TO:0002620): An epicarp anatomy and morphology trait (TO:0000947) that is associated with the qualities of the epicarp (PO:xxx).
mesocarp color (TO:0002619): An mesocarp anatomy and morphology trait (TO:0000948) that is associated with the qualities of the mesocarp (PO:0009087).

added Xrefs to:
endocarp color (TO:0002618): PATO:0000014
epicarp color (TO:0002620): PATO:0000014
mesocarp color (TO:0002619): PATO:0000014