TO Edits: Added Xrefs to GitHub/ SourceForge trackers: TO_GIT: 248 - plant growth and development trait (TO:0000357)
TO_GIT: 252
- plant structure growth and development trait (TO:0000928)

plant embryo development trait (TO:0000620):
- Was: is_a plant structure growth and development trait (TO:0000928)- moved to  is_a whole plant growth and development trait (TO:0000947)

Revised name and definiton:

was: embryo development trait (TO:0000620): The traits that are assayed during embryogenesis starting from the zygote to the germination phase.

New:
plant embryo development trait (TO:0000620): A whole plant growth and development trait (TO:0000947) that is associated with a plant embryo development stage (PO:0007631).

Added Xref to TO_GIT: 299