Updated the Xrefs from the old  OBO_SF_TO: 7 digit to  OBO_SF2_TO: numerical for the following terms- plant trait TO:0000387 - fixed several errors in Xrefs
Created a couple of new terms to describe the corolla:

corolla anatomy and morphology trait (TO:0000950):
A collective phyllome structure anatomy and morphology trait (TO:0000845) which is associated with the qualities of a corolla (PO:0009059).

corolla color (TO:0000951): A corolla anatomy and morphology trait (TO:0000950) that is associated with the color of a corolla (PO:0009059). Added Xref to PATO:0000014