stem number (TO:0000953): Revised parent and def'n
Was: stem number (TO:0000953): A stem anatomy and morphology trait (TO:0000361) which is associated with the number of stems (PO:0009047).
Xrefs to: number (PATO:0001555); TO_GIT:157
narrow synonym: stalk weight

Revised:
stem number (TO:0000953): A shoot system anatomy and morphology trait (TO:0000077) which is associated with the number of stems (PO:0009047).
- removed narrow synonym: stalk weight