Created new term: ear color (TO:0000963): An infructescence anatomy and morphology trait (TO:0000920) that is associated with the color of an ear infructescence (PO:0025597). Added Xrefs: color (PATO:0000014); TO_GIT:132
Created new term: ear shape (TO:0000964): An infructescence anatomy and morphology trait (TO:0000920) that is associated with the shape of an ear infructescence (PO:0025597).
Added Xrefs: shape (PATO:0000052); TO_GIT:132

Created new term: ear aspect (TO:0000965): An infructescence anatomy and morphology trait (TO:0000920) which is associated with the qualitative rating of the harvested ear infructescence (PO:0025597).
Added Xref: TO_GIT:132

Created new term: ear position (TO:0000966): An infructescence anatomy and morphology trait (TO:0000920) which is associated with the ratio of ear height to plant height.
Added Xref: TO_GIT:132