Changes to exocarp, mesocarp, pericarp morphology terms:
exocarp anatomy and morphology trait (TO:0000906) and child terms; exocarp color (TO:0002620) and exocarp thickness (TO:0000907):
- fixed PO:id for exocarp in definitions and GitHub tracker # 295
- fixed xref to correct TO_GIT: 295

- Moved epicarp texture (TO:0002624) to be is_a exocarp anatomy and morphology trait (TO:0000906). Added Xref: TO_GIT: 295
- revised name: exocarp texture (TO:0002624)- was changed in commit bec2d852964ecfddbeb7e1a73fc4a98e75cb13c4 on July 23, 2015 but does not seem to be there.

Added missing def'n for texture -
exocarp texture (TO:0002624): An exocarp anatomy and morphology trait (TO:0000906) that is associated with the texture of the exocarp (PO:0009085).

Revised def'n:  exocarp color (TO:0002620): An exocarp anatomy and morphology trait (TO:0000906) that is associated with the color of the exocarp (PO:0009085).

Fixed xref to TO_GIT:292 (was 293):
- pericarp anatomy and morphology trait (TO:0000945)
- pericarp color (TO:0000707)
- pericarp texture (TO:0002623)

Revised def'n: mesocarp color (TO:0002619): A mesocarp anatomy and morphology trait (TO:0000948) that is associated with the color of the mesocarp (PO:0009087).
- fixed xref to correct TO_GIT: 294