pseudostem morphology traits
Revised def'ns:
- collective organ part structure anatomy and morphology trait (TO:0000841):  A collective plant structure anatomy and morphology trait (TO:0000840) which is associated with the qualities of a collective organ part structure (PO:0025269).
 TO_GIT: 150

Moved pseudostem anatomy and morphology trait  and child terms to the correct parent- a collective organ part structure anatomy and morphology trait (TO:0000841).

pseudostem anatomy and morphology trait (TO:0000807):
-  was: A stem anatomy and morphology trait (TO:0000361) which is associated with a pseudostem (PO:0025248).
Revised:  A collective organ part structure anatomy and morphology trait (TO:0000841) which is associated with the qualities of a pseudostem (PO:0025248).

Fixed xrefs to GitHub tracker on all five : should be TO_GIT: 94, not 95,