panicle density (TO:0000970): An inflorescence density (TO:0000800) trait associated with a panicle (PO:0009049). is_a  panicle anatomy and morphology trait (TO:0000847)
inflorescence height (TO:0000682):- moved to be an inflorescence anatomy and morphology trait (TO:0000373), was a height related trait (TO:0000171) .

Revised def'n: An inflorescence anatomy and morphology trait (TO:0000373)
 which is associated with the height of an inflorescence (PO:0009049) measured from ground to the attachment point.
 Added: head height (narrow); Added Xref  to TO_GIT:97

New term: inflorescence hairiness (TO:0000971): An inflorescence anatomy and morphology trait (TO:0000373) which is associated with the presence or absence of hairs on an inflorescence (PO:0009049).

Fixed duplicated stem size term which was added accidentally- deleted term stem size TO:0000954.  Moved synonym stalk size (related) to TO:0000892.