Revised names and def'ns: was: morphology trait (TO:0000017) and def'n: A plant trait (TO:0000387) which is a morphological quality of a plant anatomical entity(PO:0025131) or one or more constituent cellular components (GO:0005575)contained therein.
Revised:
plant morphology trait (TO:0000017):  A plant trait (TO:0000387) which is a morphological quality of a plant anatomical entity (PO:0025131) or a constituent cellular component (GO:0005575) contained therein.

Was: plant structure anatomy and morphology trait (TO:0000839): A morphology trait (TO:0000017) which is associated with a plant structure (PO:0009011).

Revised:
plant structure morphology trait(TO:0000839): A plant morphology trait (TO:0000017) which is a morphological quality of a plant structure (PO:0009011).