ve plant structure anatomy and morphology trait (TO:0000840): Revised name and def'n Was: plant structure anatomy and morphology trait (TO:0000839), A plant structure morphology trait (TO:0000839) which is associated with a collective plant structure (PO:0025497). New: plant structure morphology trait (TO:0000840), A plant structure morphology trait (TO:0000839) which is a quality of a collective plant structure (PO:0025497).
collective organ part structure anatomy and morphology trait (TO:0000841) : Revised name and def'n
Was: collective organ part structure anatomy and morphology trait (TO:0000841), A collective plant structure anatomy and morphology trait (TO:0000840) which is associated with a collective plant organ structure (PO:0025007).
New: collective plant organ structure morphology trait (TO:0000841), A collective plant structure morphology trait (TO:0000840) which is a quality of a collective plant organ structure (PO:0025007).

collective plant organ structure anatomy and morphology trait (TO:0000842): Revised name and def'n
Was: collective plant structure anatomy and morphology trait (TO:0000840), A collective plant structure anatomy and morphology trait (TO:0000840) which is associated with a collective plant organ structure (PO:0025007).
New: collective plant organ structure morphology trait (TO:0000842), A collective plant structure morphology trait (TO:0000840) which is a quality of a collective plant organ structure (PO:0025007).

collective phyllome structure anatomy and morphology trait (TO:0000845): revised name and def'n
Was: collective phyllome structure anatomy and morphology trait (TO:0000845), A collective plant organ structure anatomy and morphology trait (TO:0000842) which is associated with a collective phyllome structure (PO:0025023).
New: collective phyllome structure morphology trait (TO:0000845), A collective plant organ structure morphology trait (TO:0000842) which is a quality of a collective phyllome structure (PO:0025023).

shoot system anatomy and morphology trait (TO:0000077): Revised name and def'n
Was: shoot system anatomy and morphology trait (TO:0000077), A collective plant organ structure anatomy and morphology trait (TO:0000842) which is associated with a shoot system (PO:0009006).
New: shoot system morphology trait (TO:0000077), A collective plant organ structure morphology trait (TO:0000842) which is a quality of a shoot system (PO:0009006).

reproductive shoot system anatomy and morphology trait (TO:0000786): Revised name and def'n
Was: reproductive shoot system anatomy and morphology trait (TO:0000786), A shoot system anatomy and morphology trait (TO:0000077) that is associated with a reproductive shoot system (PO:0025082).
New: reproductive shoot system morphology trait (TO:0000786), A shoot system morphology trait (TO:0000077) that is a quality of a reproductive shoot system (PO:0025082).

inflorescence anatomy and morphology trait (TO:0000373):Revised name and def'n Was: inflorescence anatomy and morphology trait (TO:0000373), A reproductive shoot system anatomy and morphology trait (TO:0000786) that is associated with the qualities of an inflorescence (PO:0009049).
New: inflorescence morphology trait (TO:0000373), A reproductive shoot system morphology trait (TO:0000786) that is a quality of an inflorescence (PO:0009049).

it message for your changes. Lines starting