Edits to involucre bract terms involucral bract spine presence (TO:0020120): was a child of TO:0000499 flower anatomy and morphology trait.
Moved to be is_a  inflorescence bract morphology trait (TO:0000819).

Revised def'n: involucral bract spine presence (TO:0020120): An inflorescence bract morphology trait (TO:0000819) which is the presence of  spines on an involucral bract (PO:0009045).
Added xref to TO_GIT:110

revised def'ns:
* involucre bristle length (TO:0000623):
was: An inflorescence bristle length trait (TO:0000643) that is a quality of the bristles found on an involucre (PO:0009100).

Revised: An inflorescence bristle length trait (TO:0000643) that is the length of the bristles found on an involucre (PO:0009100).

* inflorescence bristle length (TO:0000643):
was;  An inflorescence morphology trait (TO:0000050) that is a quality of the bristles found on an inflorescence (PO:0009049).

Revised: An inflorescence morphology trait (TO:0000050) that is the length of the bristles found on an inflorescence (PO:0009049).