term name and definition revisions Revised name and def'n of: pseudostem anatomy morphology trait (TO:0000807) Was: pseudostem anatomy morphology trait (TO:0000807), A collective organ part structure anatomy and morphology trait (TO:0000841) which is associated with the qualities of a pseudostem (PO:0025248). New: pseudostem morphology trait (TO:0000807), A collective organ part structure morphology trait (TO:0000841) which is a quality of a pseudostem (PO:0025248).
Revised def'n of: psuedostem color (TO:0000810)
Was: A pseudostem anatomy and morphology trait (TO:0000807) which is associated with the color of a pseudostem (PO:0025248).
New: A pseudostem morphology trait (TO:0000807) that is the color of a pseudostem (PO:0025248).

Revised def'n of: pseudostem diameter (TO:0000811)
Was: A pseudostem anatomy and morphology trait (TO:0000807) which is associated with the diameter of a pseudostem (PO:0025248).
New: A pseudostem morphology trait (TO:0000807) that is the diameter of a pseudostem (PO:0025248).

Revised def'n of: pseudostem length (TO:0000808)
Was: A pseudostem anatomy and morphology trait (TO:0000807) which is associated with the length of a pseudostem (PO:0025248).
New: A pseudostem morphology trait (TO:0000807) that is the length of a pseudostem (PO:0025248).

Revised def'n of: pseudostem shape (TO:0000809)
Was: A pseudostem anatomy and morphology trait (TO:0000807) which is associated with the shape of a pseudostem (PO:0025248).
New: A pseudostem morphology trait (TO:0000807) that is the shape of a pseudostem (PO:0025248).