Revised: perianth morphology trait (TO:0000275) #306
Was: perianth anatomy and morphology trait (TO:0000275): A collective phyllome structure anatomy and morphology trait (TO:0000845) which is associated with the qualities of a perianth (PO:0009058).

Revised name and definition: perianth morphology trait (TO:0000275):
A collective phyllome structure morphology trait (TO:0000845) which is a quality of a perianth (PO:0009058).

Also various minor edits