Revisions to term names and definitions: Revised name and def'n: corolla color (TO:0000951) Was: corolla color (TO:0000951), A corolla anatomy and morphology trait (TO:0000950) that is associated with the color of a corolla (PO:0009059). New: corolla color (TO:0000951),A corolla morphology trait (TO:0000950) that is the color of a corolla (PO:0009059).
Revised name and def'n: cardinal part of a multi-tissue plant structure anatomy and morphology trait (TO:0000837)
Was: cardinal part of a multi-tissue plant structure anatomy and morphology trait (TO:0000837), A plant structure anatomy and morphology trait (TO:0000839) which is associated with a cardinal organ part of a multi-tissue plant structure (PO:0025498).
New: cardinal part of a multi-tissue plant structure morphology trait (TO:0000837), A plant structure morphology trait (TO:0000839) which is a quality of a cardinal organ part of a multi-tissue plant structure (PO:0025498).

Revised name and def'n: cardinal organ part anatomy and morphology trait (TO:0000754)
Was: cardinal organ part anatomy and morphology trait (TO:0000754), A cardinal organ part of a multi-tissue plant structure anatomy and morphology trait (TO:0000754) which is associated with a cardinal organ part (TO:0000754).
New: cardinal organ part morphology trait (TO:0000754), A cardinal organ part of a multi-tissue plant structure morphology trait (TO:0000837) which is a quality of a cardinal organ part (PO:0025001).

Revised name and def'n: in vitro plant structure anatomy and morphology trait (TO:0000844).
Was: in vitro plant structure anatomy and morphology trait (TO:0000844), A plant structure anatomy and morphology trait (TO:0000839) which is associated with an in vitro plant structure.
New: in vitro plant structure morphology trait (TO:00008440), A plant structure morphology trait (TO:0000839) which is a quality of an in vitro plant structure (PO:0000004).

Revised name and def'n: multi-tissue plant structure anatomy and morphology trait (TO:0000836).
Was: multi-tissue plant structure anatomy and morphology trait (TO:0000836), A plant structure anatomy and morphology trait (TO:0000839) which is associated with a multi-tissue plant structure (PO:0025496).
New: multi-tissue plant structure morphology trait (TO:0000836), A plant structure morphology trait (TO:0000839) which is a quality of a multi-tissue plant structure (PO:0025496).

Revised name and def'n: plant cell anatomy and morphology trait (TO:0002683)
Was: plant cell anatomy and morphology trait (TO:0002683), A plant structure anatomy and morphology trait (TO:0000839) which is associated with changes in the anatomy or structure of the plant cell (PO:0009002) and its parts.
New: plant cell morphology trait (TO:0002683), A plant structure morphology trait (TO:0000839) which is a quality of a plant cell (PO:0009002).
added xref (TO_GIT: 307)

Revised name and def'n: plant embryo anatomy and morphology trait (TO:0000064)
Was: plant embryo anatomy and morphology trait (TO:0000064), A plant structure anatomy and morphology trait (TO:0000839) which is associated with a plant embryo (PO:0009009).
New: plant embryo morphology trait (TO:0000064), A plant structure morphology trait (TO:0000839) which is a quality of a plant embryo (PO:0009009).

Revised name and def'n: portion of plant tissue anatomy and morphology trait (TO:0000843)
Was: portion of plant tissue anatomy and morphology trait (TO:0000843), A plant structure anatomy and morphology trait (TO:0000839) which is associated with a portion of plant tissue (PO:0009007).
New: portion of plant tissue morphology trait (TO:0000843), A plant structure morphology trait (TO:0000839) which is a quality of a portion of plant tissue (PO:0009007).

Revised name and def'n: trichome anatomy and morphology trait (TO:0000911)
Was: trichome anatomy and morphology trait (TO:0000911), A plant structure anatomy and morphology trait (TO:0000839) which is associated with a trichome (PO:0000282).
New: trichome morphology trait (TO:0000911), A plant structure morphology trait (TO:0000839) which is a quality of a trichome (PO:0000282).
Revised xref from (TO_GIT:227) to (TO_GIT:228)